Implemented password encryption

+++ server/src/main/java/service/AuthService.java
@@ -6,7 +6,15 @@ import dataobjects.*;
 import java.util.UUID;
 
 public class AuthService {
-    private static final AuthDAO AUTH_DATA_ACCESS = new MemoryAuthDAO();
+    private static final AuthDAO AUTH_DATA_ACCESS;
+
+    static {
+        try {
+            AUTH_DATA_ACCESS = new SQLAuthDAO();
+        } catch (DataAccessException e) {
+            throw new RuntimeException(e);
+        }
+    }
 
     /**
      * clears everything in the auth data database
@@ -31,12 +39,12 @@ public class AuthService {
      * @param authToken the auth token to remove from the database
      * @throws UnrecognizedAuthTokenException if the authToken is null or not in the database
      */
-    public static void logoutUser(String authToken) throws UnrecognizedAuthTokenException {
-        try {
-            AuthData auth = AUTH_DATA_ACCESS.getAuth(authToken);
-            AUTH_DATA_ACCESS.deleteAuth(auth);
-        } catch (DataAccessException | NullPointerException e) {
+    public static void logoutUser(String authToken) throws UnrecognizedAuthTokenException, DataAccessException {
+        AuthData auth = AUTH_DATA_ACCESS.getAuth(authToken);
+        if(auth == null) {
             throw new UnrecognizedAuthTokenException("Unrecognized Auth Token: " + authToken);
+        } else {
+            AUTH_DATA_ACCESS.deleteAuth(auth);
         }
     }
 
@@ -45,10 +53,8 @@ public class AuthService {
      * @param authToken the auth token to validate
      * @throws UnrecognizedAuthTokenException if the auth token isn't in the database
      */
-    public static void validateAuth (String authToken) throws UnrecognizedAuthTokenException {
-        try {
-            AUTH_DATA_ACCESS.getAuth(authToken);
-        } catch (DataAccessException _) {
+    public static void validateAuth (String authToken) throws UnrecognizedAuthTokenException, DataAccessException {
+        if(AUTH_DATA_ACCESS.getAuth(authToken) == null) {
             throw new UnrecognizedAuthTokenException("Invalid Auth Token");
         }
     }
